+++ src/ua/data_types/server_state.rs
@@ -1,0 +1,1 @@
+crate::data_type!(ServerState);